Add files via upload
Field of view is being cut off at the back of the head. There is also an area of signal dropout due to surface coil sensitivity. This means the back of the head was too close to the receiver. This can be reduced by ensuring there is sufficient padding at the back of the head.